+++ docs/CHANGES.rst
@@ -4,11 +4,11 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- I18n fixes.
-  [vincentfretin]
-
 - Don't batch the already batched collection results. Fixes #221.
   [thet]
+
+- I18n fixes.
+  [vincentfretin]
 
 - Update image_view_fullscreen.pt for mobile friendliness.
   [fulv]