+++ docs/CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.2a8 (unreleased)
 ------------------
 
+- Fix ``contentFilter`` for collections.
+  [thet]
+
 - Cache the results of the collection and folder view on the request. The
   results might be called again by other packages to extract information on the
   result list.